Big Time Plus watch face
Enhancements to Big Time.
a) hour clock is left aligned for single digit. this was very irritating and I think this is way more readable.
a) Inspired by apple iphone clock, AM is shown in inverted black on white, PM time is shown in white on black as the default.
c) units_changed checked to only update hour or minute